% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioTIP-package.R
\docType{package}
\name{BioTIP-package}
\alias{BioTIP}
\alias{BioTIP-package}
\title{BioTIP: BioTIP: An R package for characterization of Biological Tipping-Point}
\description{
Adopting tipping-point theory to transcriptome profiles 
     to unravel disease regulatory trajectory.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/xyang2uchicago/BioTIP}
}

}
\keyword{internal}
