% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cod}
\alias{cod}
\title{cod dataset}
\format{A dataframe with 3 data columns and 1 metadata column.
\describe{
\item{seqnames}{chromosome names ex.chr21}
\item{start}{a numeric value of genome starting site}
\item{end}{a numeric value of genome ending site}
\item{width}{a numeric value of end-start+1}
\item{strand}{specific strand of the genomic location (+,-,*)}
\item{biotype}{biotype obtained through gtf file}
}}
\usage{
cod
}
\description{
A subset of gencode_gr extracted as:
cod <- subset(gencode_gr, biotype == 'protein_coding')
}
\keyword{datasets}
