% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gencode}
\alias{gencode}
\title{chr21 reference genome from GENCODE GRCh37}
\format{A GRanges object of chr21 with 28082 rows and 4 variables:
\describe{
\item{senames}{chromosome names ex.chr21}
\item{ranges}{genomic ranges on the genome}
\item{strand}{specific strand of the genomic location ex.+,-,*}
\item{biotype}{biotype obtained through gtf file}
}}
\usage{
gencode
}
\description{
A reference human GRCh37 genome dataset from GENCODE.
\url{https://www.gencodegenes.org/human/release_25lift37.html}. Included
variables are as follows:
}
\keyword{datasets}
