% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioTIP.R
\name{getCI_inner}
\alias{getCI_inner}
\title{calculate MCI score}
\usage{
getCI_inner(members, countsL, adjust.size)
}
\value{
a list MCI scores
}
\description{
calculate MCI score for randomly selected loci. This function will be called by simulationMCI
}
\author{
Zhezhen Wang
}
