% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPS.R
\name{getCluster}
\alias{getCluster}
\title{getCluster}
\usage{
getCluster(igraphL, steps = 4)
}
\arguments{
\item{igraphL}{a list of count matrix or a list of igraph object}

\item{steps}{number of steps}
}
\value{
A list of objects
}
\description{
getCluster
}
\examples{
getCluster(igraphL, steps = 4)

}
\author{
Zhezhen Wang
}
