% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPS.R
\name{getMaxCImember}
\alias{getMaxCImember}
\title{get the index and members with the maximum CI score}
\usage{
getMaxCImember(membersL, CIl, minsize = 1)
}
\arguments{
\item{membersL}{a list of characters, the first element of output from function getCI}

\item{CIl}{alist of numerics, the secned element of output from function getCI}

\item{minsize}{a numeric value of minimum size allowed for a cluster}
}
\value{
a list of index and members
}
\description{
get the index and members with the maximum CI score
}
\examples{
membersL_noweight <- getCI(cluster,test,adjust.size = F,ylim = c(0,8))
maxCIms <- getMaxCImember(membersL_noweight[[1]],membersL_noweight[[2]],min =3)
head(maxCIms)

}
\author{
Zhezhen Wang
}
