% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE6136_matrix}
\alias{GSE6136_matrix}
\title{GSE6136 matrix dataset}
\format{
A dataframe with 22690 columns and 27 rows column.
\describe{
\item{GSM142398-GSM142423}{Names of gene interest}
\item{ID_REF}{Reference ID of the target genes}
}
}
\source{
\url{https://www.gencodegenes.org/human/}
}
\usage{
GSE6136_matrix
}
\description{
A gene annotation samples with their corresponding geneId extracted from GENCODE database.
A python script was then run to extcat GSE6136_cli dataset. The script is included in the R/data_raw folder.
}
\keyword{datasets}
