% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ILEF_gr}
\alias{ILEF_gr}
\title{Chromosome ranges of chr21 dataset}
\format{A data frame of chr21 with 137 rows and 4 variables:
\describe{
\item{seqnames}{chromosome names (chr1,chr6,chr21)}
\item{ranges}{chromosome intervals ranges on the
genome(167684657--167729508,3876218--3897070)}
\item{strand}{specific strand of the genomic location (+,-,*}
\item{row names}{name of the data rows(A1BG,vawser)}
}}
\usage{
ILEF_gr
}
\description{
A dataset containing chromosomes in the genome regions of interest for 137
chromosome. The variables are as follows:
}
\keyword{datasets}
