% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intron}
\alias{intron}
\title{one example of intron in chr21}
\format{A GRanges object with 659327 rows and 5 variables:
\describe{
\item{seqnames}{chromosome names ex.chr21}
\item{ranges}{genomic ranges on the genome}
\item{strand}{specific strand of the genomic location ex.+,-,*}
}}
\usage{
intron
}
\description{
A dataset containing chromosomes in the genome regions of interest.
The variables are as follows:
}
\keyword{datasets}
