% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intron_gr}
\alias{intron_gr}
\title{Coding transcriptome in chr21 dataset}
\format{A data frame with 659327 rows and 5 variables:
\describe{
\item{seqnames}{chromosome names (chr1,chrM,chr21)}
\item{ranges}{chromosome ranges on the genome(167684657--167729508)}
\item{strand}{specific strand of the genomic location (+,-,*)}
\item{name}{internal assigned names(uc001aaa.3_intron_0_0_chr1_12228_f)}
\item{score}{score not used in this data set(0)}
}}
\usage{
intron_gr
}
\description{
A dataset containing chromosomes in the genome regions of interest.
The variables are as follows:
}
\keyword{datasets}
