% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioTIP.R
\name{optimize.sd_selection}
\alias{optimize.sd_selection}
\title{optimization of sd selection}
\usage{
optimize.sd_selection(df, samplesL, B = 100, percent = 0.8,
  times = 80, cutoff = 0.01, method = "other", control_df = NULL,
  control_samplesL = NULL)
}
\arguments{
\item{df}{A dataframe of numerics, factor or character.The rows and columns
represent unique transcript IDs (geneID) and sample names, respectively.}

\item{samplesL}{An R list of n vectors, where n equals to the number of
states. Each vector gives the sample names in a state, therefore most
possibly to be a vector of character or integer. Note that the vectors
(sample names) has to be among the column names of the R object 'df'.}

\item{cutoff}{A positive numeric value. Default is 0.01, if < 1, automaticlly
goes to select top x# transcripts using the a selecting method (which is
either the ?reference?, ?other? stages or ?pervious?# stage), e.g. by
default it will select top 1\% of the transcripts.}

\item{method}{Selection of methods from 'reference', 'other', 'previous',
'itself', or ?longitudinal reference?. Some specific requirements for each
option:
\itemize{
\item 'reference', the reference has to be the first
\item 'previous', make sure sampleL is in the right order from benign to malign
\item 'itself', make sure the cutoff is smaller than 1
\item 'longitudinal reference', make sure control_df and control_samplesL are not NULL
\item default uses 'other'
}}

\item{control_df}{a count matrix with unique loci row names:ID X column
names:samples names of control samples, required for method 'longitudinal
reference'}

\item{control_samplesL}{a list of characters with stages as names of control
samples, required for method 'longitudinal reference'}
}
\value{
A list of dataframe of filtered transcripts with highest standard
deviation are selected from \code{df} based on a cutoff value assigned. The
resulting dataframe represents a subset of the raw input \code{df}.
}
\description{
The \code{sd_selection.simulation} filter a multi-state dataset based on a cutoff value for standard deviation per state and optimize. By default, a cutoff value of 0.01 is used. Suggested if each state contains more than 10 samples.
}
\examples{

counts = matrix(sample(1:100,30),2,30)
colnames(counts) = 1:30
row.names(counts) = paste0('loci',1:2)
cli = cbind(1:30,rep(c('state1','state2','state3'),each = 10))
colnames(cli) = c('samples','group')
samplesL <- split(cli[,1],f = cli[,'group'])
test_sd_selection <- optimize.sd_selection(counts, samplesL, B = 3, cutoff =0.01)
}
\seealso{
\code{\link{sd_selection}}
}
\author{
Zhezhen Wang \email{zhezhen@uchicago.edu}
}
