% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioTIP.R
\name{plot_simulation_sample}
\alias{plot_simulation_sample}
\title{calculating Ic scores based on simulated samples}
\usage{
plot_simulation_sample(counts, sampleNo, Ic, genes, B = 1000,
  main = "simulation of samples")
}
\arguments{
\item{counts}{A  numeric matrix or data frame. The rows and columns represent unique transcript IDs (geneID) and sample names, respectively.}

\item{sampleNo}{An integer of sample size of the tipping point}

\item{Ic}{A numeric value. Ic score of identified BioTiP}

\item{genes}{A character vector of identified BioTiP unique ids}

\item{B}{An integer indicating number of times to run this simulation, default 1000.}

\item{main}{A character vector. The title of the plot. Defualt is NULL.}
}
\value{
Return a density plot of simulated Ic score with p-value
}
\description{
simulate \code{x} samples \code{B} times to calculate the Ic score, where x should be the same as the length of identified BioTiP and B is self-defined.
}
\examples{
counts = matrix(sample(1:100,27),3,9)
colnames(counts) = 1:9
row.names(counts) = c('loci1','loci2','loci3')
BioTiP = c('loci1','loci2')
plot_simulation_sample(counts,3,3.4,BioTiP,B=3)
}
\author{
Zhezhen Wang \email{zhezhen@uchicago.edu}
}
