% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioTIP.R
\name{sd_selection}
\alias{sd_selection}
\title{Selecting Highly Oscillating Transcripts}
\usage{
sd_selection(df, samplesL, cutoff = 0.01, method = "other")
}
\arguments{
\item{df}{A numeric matrix or data frame. The rows and columns represent
unique transcript IDs (geneID) and sample names, respectively.}

\item{samplesL}{A list of vectors, whose length is the number of states. Each
vector gives the sample names in a state. Note that the vectors (sample names) has
to be among the column names of the R object 'df'.}

\item{cutoff}{A positive numeric value. Default is 0.01. If < 1,
automatically selects top x  transcripts using the a selecting
method (which is either the \code{reference}, \code{other} stages or \code{previous}
stage), e.g. by default it will select top 1\% of the transcripts.}
}
\value{
\code{sd_selection()} A list of data frames, whose length is the number
of states. The rows in each data frame are the filtered transcripts with
highest standard deviation selected from \code{df} and based on an assigned cutoff
value. Each resulting data frame represents a subset of the raw
input \code{df}, with the sample ID of the same state in the column.
}
\description{
\code{sd_selection} pre-selects highly oscillating transcripts
from the input dataset \code{df}. The dataset must contain multiple sample
groups (or 'states'). For each state, the function filters the dataset using
a cutoff value for standard deviation. The default cutoff value is 0.01
(i.e., higher than the top 1\% standard deviation).
}
\examples{
counts = matrix(sample(1:100,18),2,9)
colnames(counts) = 1:9
row.names(counts) = c('loci1','loci2')
cli = cbind(1:9,rep(c('state1','state2','state3'),each = 3))
colnames(cli) = c('samples','group')
samplesL <- split(cli[,1],f = cli[,'group'])
test_sd_selection <- sd_selection(counts, samplesL, 0.01)

}
\seealso{
\code{\link{sd_selection.simulation}}
}
\author{
Zhezhen Wang \email{zhezhen@uchicago.edu}
}
