% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPS.R
\name{getNetwork}
\alias{getNetwork}
\title{get an igraph object based on Pearson Correlation Coefficiency(PCC)}
\usage{
getNetwork(optimal, p = 0.05)
}
\arguments{
\item{optimal}{a list of count matrix}

\item{p}{a numeric cutoff}
}
\value{
a list of an igraph object
}
\description{
get a igraph object based on Pearson Correlation Coefficiency(PCC)
}
\examples{
igraphL <- getNetwork(test, p=1)

}
\author{
Zhezhen Wang and Biniam Feleke
}
