% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE6136_cli}
\alias{GSE6136_cli}
\title{GSE6136 cli dataset}
\format{A dataframe with 36 columns and 27 rows column.
\describe{
\item{Columns}{Sample names or IDs}
\item{Rows}{Clinical features of the samples}
}}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE6136}
}
\usage{
GSE6136_cli
}
\description{
A gene annotation samples with their corresponding geneId extracted from GENCODE database.
A python script was then run to extcat GSE6136_cli dataset. The script is included in the R/data_raw folder.
}
\keyword{datasets}
