% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioTIP.R
\name{plot_Ic_Simulation}
\alias{plot_Ic_Simulation}
\title{line plot of Ic score and simulated Ic scores}
\usage{
plot_Ic_Simulation(Ic, simulation, las = 0, ylim = NULL,
  order = NULL, main = NULL)
}
\arguments{
\item{Ic}{A vector with names of states. If order is not assigned, then plot by the order of this vector.}

\item{simulation}{A numeric matrix of Ic scores in which rows are states and columns are numbers of simulated times. It can be obtained from \code{\link{simulation_loci}}}

\item{las}{Numeric in {0,1,2,3}; the style of axis labels. Default is 0, meaning labels are parallel. (link to http://127.0.0.1:21580/library/graphics/html/par.html)}

\item{ylim}{An integer vector of length 2. Default is NULL.}

\item{order}{A vector of state names in the customized order to be plotted, set to NULL by default.}

\item{main}{A character vector. The title of the plot. Defualt is NULL.}
}
\description{
generate a line plot of Ic score and simulated Ic scores.
}
\examples{
sim = matrix(sample(1:10,9),3,3)
row.names(sim) = paste0('state',1:3)
Ic = c('state1' = 3.4,'state2' = 5.6,'state3' = 2)
plot_Ic_Simulation(Ic,sim)
}
\author{
Zhezhen Wang \email{zhezhen@uchicago.edu}
}
