% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioTIP.R
\name{plot_MCI_Simulation}
\alias{plot_MCI_Simulation}
\title{Simulation of Loci to Calculate the CI Score}
\usage{
plot_MCI_Simulation(MCI, simulation, las = 0, order = NULL,
  ylim = NULL, main = NULL)
}
\arguments{
\item{MCI}{A named vector of max CI scores per state, can be obtained from function \code{\link{getMaxStats}}}

\item{simulation}{A matrix state * number of simulated times, can be obtained from function \code{\link{simulationMCI}}}

\item{las}{Numeric in {0,1,2,3}; the style of axis labels. Default is 0, meaning labels are parallel. (link to http://127.0.0.1:21580/library/graphics/html/par.html)}

\item{order}{A vector of state names in the customized order to be plotted, set to NULL by default.}

\item{ylim}{An integer vector of length 2. Default is NULL.}

\item{main}{A character vector. The title of the plot. Defualt is NULL.}
}
\description{
simulation of loci to calculate the Ic score.
}
\examples{
MCI = c(1:3); names(MCI) = c('a','b','c')
simMCI = matrix(sample(1:100,9),3,3)
row.names(simMCI) = names(MCI)
plot_MCI_Simulation(MCI,simMCI)

}
\author{
Zhezhen Wang \email{zhezhen@uchicago.edu}
}
