% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioTIP.R
\name{plotIc}
\alias{plotIc}
\title{plot a line plot of Ic score for each state}
\usage{
plotIc(Ic, las = 0, order = NULL)
}
\arguments{
\item{Ic}{A vector with names of states. if order is not assigned then plot by the order of this vector}

\item{las}{Numeric in {0,1,2,3}; the style of axis labels. Default is, meaning labels are parallel (=0). (link to http://127.0.0.1:21580/library/graphics/html/par.html)}

\item{order}{A vector of state names in the customalized order to be plotted, setting to NULL by default.}
}
\description{
plot a line plot with Ic score for each state
}
\examples{
Ic = c('state3' = 3.4,'state1' = 5.6,'state2' = 2)
plotIc(Ic,order = c('state1','state2','state3'))
}
\author{
Zhezhen Wang \email{zhezhen@uchicago.edu}
}
